fix(docs): move footer to app to fix alignment of footer
