+++ packages/apps/docs/src/components/Footer/FooterText.tsx
